Rename plot.sh to test.sh and correct a small bug in the script.
When the -w parameter was passed a value of 3 the average was computed
taking off the best and worst result, thus making the average on just 1
execution. A more meaningful behavior is to compute the average on all 3
executions, otherwise it would be as the executions were just 1, but the time
spent to complete the script is 3 times higher.